[blur-view] Move `useResettingState` to the upper level utils & refactor `useLoopingAnimatedValue`
